error fixed

+++ tasks/seq/ivanchenko_a_gauss_filter_vertical/src/ops_seq.cpp
@@ -41,8 +41,8 @@ void GaussFilterSequential::createKernel(float sigma) {
   uint32_t radius = kernelSize / 2;
   uint32_t size = kernelSize;
   float norm = 0;  // коэффициент нормировки ядра
-  for (int i = -radius; i <= (int)radius; i++) {
-    for (int j = -radius; j <= (int)radius; j++) {
+  for (int i = -(int)radius; i <= (int)radius; i++) {
+    for (int j = -(int)radius; j <= (int)radius; j++) {
       kernel[i + radius][j + radius] = (double)(exp(-(i * i + j * j) / (2 * sigma * sigma)));
       norm += kernel[i + radius][j + radius];
     }
@@ -64,8 +64,8 @@ void GaussFilterSequential::applyKernel() {
 Color GaussFilterSequential::calculateNewPixelColor(size_t x, size_t y) {
   uint32_t radius = kernelSize / 2;
   float resultR = 0, resultG = 0, resultB = 0;
-  for (int l = -radius; l <= (int)radius; l++) {
-    for (int k = -radius; k <= (int)radius; k++) {
+  for (int l = -((int)radius); l <= (int)radius; l++) {
+    for (int k = -((int)radius); k <= (int)radius; k++) {
       size_t idX = x + k, idY = y + l;
       if (idX < 0) idX = 0;
       if (idX >= width) idX = width - 1;
